Protocol Version is a Byte, not a Int

+++ proxy/src/main/java/net/md_5/bungee/connection/InitialHandler.java
@@ -161,7 +161,7 @@ public class InitialHandler extends PacketHandler implements PendingConnection
             forced.ping( pingBack );
         } else
         {
-            pingBack.done( new ServerPing( -1, "", motd, bungee.getOnlineCount(), listener.getMaxPlayers() ), null );
+            pingBack.done( new ServerPing( (byte) -1, "", motd, bungee.getOnlineCount(), listener.getMaxPlayers() ), null );
         }
     }
